stock_prediction/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
torch>=1.9.0
tqdm>=4.62.0
pyyaml>=6.0